+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Todas as alterações notáveis ao projeto devem ser documentadas neste arquivo.
 O formato é baseado em [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) e este projeto é aderente ao versionamento semântico [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+  - NPM Jest Styled Components
+  - NPM enzyme-to-json
+  - Add more test to Button component
+### Changes
+  - Configs to not call styled component with prefixed filename in tests
+
 ### [0.4.1-beta]
 ### Fix
   - Decrease z-index of table loader